MLToolkit.Plots

MLToolkit.Plots.FeatureActivations — Type

A plot of feature activations.

MLToolkit.Plots.ImageGrid — Type

A plot of images in grid.

MLToolkit.Plots.LinesWithErrorBar — Type

A plot of lines with shaded error bar.

MLToolkit.Plots.OneDimFunction — Type

A plot of 1D function.

MLToolkit.Plots.TwoDimFunction — Type

A plot of 2D function.

MLToolkit.Plots.TwoYAxesLines — Type

A plot of two lines with a shared y-axis.

FileIO.save — Function

save(fn::String, p::AbstractPlot, p=nothing; bbox_inches="tight", kwargs...)

Usage:

fig = plot(p)
save("fig.png", fig, p)
save("fig.tex", fig, p)

MLToolkit.Plots.get_tikz_code — Function

get_tikz_code([fig], p::AbstractPlot; kwargs...)

Usage:

fig = plot(p)
code = get_tikz_code(fig, p)

MLToolkit.Plots.make_imggrid — Method

make_imggrid(img::AbstractArray{<:Number,4}, nrows, ncols; gap::Int=1)

Create a nrows by ncols image grid.

MLToolkit.Plots.make_imggrid — Method

make_imggrid(img; kwargs...)

Create an approximately squared image grid based on the number of images. For example, if last(size(img)) is 100, the grid will be 10 by 10.

MLToolkit.Plots.plot! — Method

plot!([ax=plt.gca()], p::AbstractPlot, args...; kwargs...)

Usage:

fig, ax = figure()
plot!(ax, p)

MLToolkit.Plots.plot — Method

plot(p::AbstractPlot, args...; figsize=nothing, kwargs...)

Usage:

fig = plot(p)